+++ src/parser/mod.rs
@@ -1,8 +1,11 @@
 mod internal_ast;
+mod internal_generator;
 mod internal_parser;
 mod sql_parser;
 mod text;
 
+pub(crate) use internal_generator::generate_internal;
+pub(crate) use internal_parser::parse_internal;
 pub use sql_parser::Parser;
 
 use crate::ast::Statement;